requirements.txt
# run this command to install all packages: pip install -r requirements.txt
python-telegram-bot
openai


